requirements_api.txt
# Official OpenAI API client for ChatGPT integration
openai==1.51.0

# JSON Schema validation supporting Draft 2020-12 for contract data validation
jsonschema==4.17.0

# Extract text content from Microsoft Word .docx files
python-docx==1.1.0

# Extract text content from PDF files for contract analysis
pdfminer.six==20221105

# Generate PDF reports and output documents
reportlab==4.0.0

# Simple cross-platform GUI with drag & drop functionality for Windows
PySimpleGUI==5.0.8.3

# FastAPI web framework for building REST APIs
fastapi==0.104.1

# ASGI server for production deployment
uvicorn[standard]==0.24.0

# API documentation and security
python-multipart==0.0.6

# HTTP client for API requests
httpx==0.25.2

# Environment variable management
python-dotenv==1.0.0
